+++ src/devices/thermostats.ts
@@ -100,7 +100,7 @@ export class Thermostats {
 
     // Set Min and Max
     if (device.changeableValues.heatCoolMode === 'Heat') {
-      this.platform.log.debug('Thermostat %s -', this.accessory.displayName, 'Device is in "Heat" mode');
+      this.platform.debug(`Thermostat ${this.accessory.displayName}, is in "${device.changeableValues.heatCoolMode}" mode`);
       this.service
         .getCharacteristic(this.platform.Characteristic.TargetTemperature)
         .setProps({
@@ -112,7 +112,7 @@ export class Thermostats {
           return this.TargetTemperature!;
         });
     } else {
-      this.platform.log.debug('Thermostat %s -', this.accessory.displayName, 'Device is in "Cool" mode');
+      this.platform.debug(`Thermostat ${this.accessory.displayName}, is in "${device.changeableValues.heatCoolMode}" mode`);
       this.service
         .getCharacteristic(this.platform.Characteristic.TargetTemperature)
         .setProps({
@@ -150,24 +150,19 @@ export class Thermostats {
 
     // Fan Controls
     if (this.platform.config.options?.thermostat?.hide_fan) {
-      if (this.platform.debugMode) {
+      if (this.platform.config.options.debug) {
         this.platform.log.error('Removing service');
       }
       this.fanService = this.accessory.getService(this.platform.Service.Fanv2);
       accessory.removeService(this.fanService!);
     } else if (!this.fanService && device.settings?.fan) {
-      if (this.platform.debugMode) {
+      if (this.platform.config.options?.debug) {
         this.platform.log.warn('Adding service');
       }
-      this.platform.log.debug(
-        'Thermostat %s -',
-        this.accessory.displayName,
-        'Available FAN settings',
-        JSON.stringify(device.settings.fan),
-      );
+      this.platform.debug(`Thermostat ${this.accessory.displayName}, Available Fan Settings ${JSON.stringify(device.settings.fan)}`);
       (this.fanService =
         this.accessory.getService(this.platform.Service.Fanv2) ||
-        this.accessory.addService(this.platform.Service.Fanv2)), '%s %s Fan', device.name, device.deviceClass;
+        this.accessory.addService(this.platform.Service.Fanv2)), `${device.name} ${device.deviceClass} Fan`;
 
       this.fanService
         .getCharacteristic(this.platform.Characteristic.Active)
@@ -177,25 +172,25 @@ export class Thermostats {
         .getCharacteristic(this.platform.Characteristic.TargetFanState)
         .onSet(this.setTargetFanState.bind(this));
     } else {
-      if (this.platform.debugMode) {
+      if (this.platform.config.options?.debug) {
         this.platform.log.warn('Fanv2 not added.');
       }
     }
 
     // Humidity Sensor Service
     if (this.platform.config.options?.thermostat?.hide_humidity) {
-      if (this.platform.debugMode) {
+      if (this.platform.config.options.debug) {
         this.platform.log.error('Removing service');
       }
       this.humidityService = this.accessory.getService(this.platform.Service.HumiditySensor);
       accessory.removeService(this.humidityService!);
     } else if (!this.humidityService && device.indoorHumidity) {
-      if (this.platform.debugMode) {
+      if (this.platform.config.options?.debug) {
         this.platform.log.warn('Adding service');
       }
       (this.humidityService =
         this.accessory.getService(this.platform.Service.HumiditySensor) ||
-        this.accessory.addService(this.platform.Service.HumiditySensor)), '%s %s HumiditySensor', device.name, device.deviceClass;
+        this.accessory.addService(this.platform.Service.HumiditySensor)), `${device.name} ${device.deviceClass} HumiditySensor`;
 
       this.humidityService
         .getCharacteristic(this.platform.Characteristic.CurrentRelativeHumidity)
@@ -206,7 +201,7 @@ export class Thermostats {
           return this.CurrentRelativeHumidity!;
         });
     } else {
-      if (this.platform.debugMode) {
+      if (this.platform.config.options?.debug) {
         this.platform.log.warn('HumiditySensor not added.');
       }
     }
@@ -237,7 +232,7 @@ export class Thermostats {
             await this.pushRoomChanges();
           } catch (e: any) {
             this.platform.log.error(JSON.stringify(e.message));
-            this.platform.log.debug('Thermostat %s -', this.accessory.displayName, JSON.stringify(e));
+            this.platform.debug(`Thermostat ${this.accessory.displayName} - ${JSON.stringify(e)}`);
             this.platform.refreshAccessToken();
             this.apiError(e);
           }
@@ -256,7 +251,7 @@ export class Thermostats {
           await this.pushChanges();
         } catch (e: any) {
           this.platform.log.error(JSON.stringify(e.message));
-          this.platform.log.debug('Thermostat %s -', this.accessory.displayName, JSON.stringify(e));
+          this.platform.debug(`Thermostat ${this.accessory.displayName} - ${JSON.stringify(e)}`);
           this.platform.refreshAccessToken();
           this.apiError(e);
         }
@@ -275,7 +270,7 @@ export class Thermostats {
             await this.pushFanChanges();
           } catch (e: any) {
             this.platform.log.error(JSON.stringify(e.message));
-            this.platform.log.debug('Thermostat %s -', this.accessory.displayName, JSON.stringify(e));
+            this.platform.debug(`Thermostat ${this.accessory.displayName} - ${JSON.stringify(e)}`);
             this.platform.refreshAccessToken();
             this.apiError(e);
           }
@@ -325,12 +320,8 @@ export class Thermostats {
       default:
         this.CurrentHeatingCoolingState = 0;
     }
-    this.platform.log.debug(
-      'Thermostat %s Heat -',
-      this.accessory.displayName,
-      'Device is Currently: ',
-      this.CurrentHeatingCoolingState,
-    );
+    this.platform.debug(`Thermostat ${this.accessory.displayName},
+     Currently Mode: ${this.device.operationStatus.mode}(${this.CurrentHeatingCoolingState})`);
 
     // Set the TargetTemperature value based on the current mode
     if (this.TargetHeatingCoolingState === this.platform.Characteristic.TargetHeatingCoolingState.HEAT) {
@@ -346,7 +337,7 @@ export class Thermostats {
     // Set the Target Fan State
     if (this.device.settings?.fan && !this.platform.config.options?.thermostat?.hide_fan) {
       if (this.fanMode) {
-        this.platform.log.debug('Thermostat %s Fan -', this.accessory.displayName, JSON.stringify(this.fanMode));
+        this.platform.debug(`Thermostat ${this.accessory.displayName}, Fan: ${JSON.stringify(this.fanMode)}`);
         if (this.fanMode.mode === 'Auto') {
           this.TargetFanState = this.platform.Characteristic.TargetFanState.AUTO;
           this.Active = this.platform.Characteristic.Active.INACTIVE;
@@ -373,14 +364,8 @@ export class Thermostats {
           },
         })
       ).data;
-      this.platform.log.debug(
-        'Thermostat %s -',
-        this.accessory.displayName,
-        'Fetched update for',
-        this.device.name,
-        'from Honeywell API:',
-        JSON.stringify(this.device.changeableValues),
-      );
+      this.platform.debug(`Thermostat ${this.accessory.displayName}, Fetched update for ${this.device.name}
+       from Honeywell API: ${JSON.stringify(this.device.changeableValues)}`);
       await this.refreshRoomPriority();
       if (this.device.settings?.fan && !this.platform.config.options?.thermostat?.hide_fan) {
         this.fanMode = (
@@ -390,14 +375,8 @@ export class Thermostats {
             },
           })
         ).data;
-        this.platform.log.debug(
-          'Thermostat %s Fan -',
-          this.accessory.displayName,
-          'Fetched update for',
-          this.device.name,
-          'from Honeywell Fan API:',
-          JSON.stringify(this.fanMode),
-        );
+        this.platform.debug(`Thermostat ${this.accessory.displayName}, Fan Fetched update for ${this.device.name}
+        from Honeywell Fan API: ${JSON.stringify(this.fanMode)}`);
       }
       this.parseStatus();
       this.updateHomeKitCharacteristics();
@@ -406,8 +385,8 @@ export class Thermostats {
         'Thermostat - Failed to update status of',
         this.device.name,
         JSON.stringify(e.message),
-        this.platform.log.debug('Thermostat %s -', this.accessory.displayName, JSON.stringify(e)),
       );
+      this.platform.debug(`Thermostat ${this.accessory.displayName} - ${JSON.stringify(e)}`);
       this.platform.refreshAccessToken();
       this.apiError(e);
     }
@@ -422,11 +401,7 @@ export class Thermostats {
           },
         })
       ).data;
-      this.platform.log.debug(
-        'Thermostat %s priority -',
-        this.accessory.displayName,
-        JSON.stringify(this.roompriority),
-      );
+      this.platform.debug(`Thermostat ${this.accessory.displayName}, Priority: ${JSON.stringify(this.roompriority)}`);
     }
   }
 
@@ -447,35 +422,37 @@ export class Thermostats {
     // Only include thermostatSetpointStatus on certain models
     switch (this.device.deviceModel) {
       case 'Round':
-        this.platform.log.debug('thermostatSetpointStatus not sent for Round Thermostats');
-        this.platform.log.debug(this.device.deviceModel);
+        this.platform.debug(`${this.device.deviceModel} Thermostats do not send thermostatSetpointStatus`);
         break;
       default:
         payload.thermostatSetpointStatus = this.platform.config.options?.thermostat?.thermostatSetpointStatus;
-        this.platform.log.debug('Send thermostatSetpointStatus');
-        this.platform.log.debug(this.device.deviceModel);
+        this.platform.debug(`Send thermostatSetpointStatus Model: ${this.device.deviceModel}`);
     }
 
-    // Always set autoChangeoverActive to 'true' for Round Thermostats
     switch (this.device.deviceModel) {
       case 'Round':
       case 'D6':
-        if (this.platform.debugMode) {
+        if (this.platform.config.options?.debug) {
           this.platform.log.warn('Round/D6 set autoChangeoverActive');
           this.platform.log.warn(this.device.deviceModel);
         }
-        payload.autoChangeoverActive = true;
+        // for Round  the 'Auto' feature is enabled via the special mode so only flip this bit when
+        // the heating/cooling state is set to  `Auto
+        if (this.TargetHeatingCoolingState === this.platform.Characteristic.TargetHeatingCoolingState.AUTO) {
+          this.platform.log.debug(`Heating/Cooling state set to Auto for ${this.device.deviceModel}, Force autoChangeoverActive`);
+          payload.autoChangeoverActive = true;
+        } else {
+          this.platform.log.debug(`Heating/cooling state not set to Auto for ${this.device.deviceModel}
+          , Using device setting ${this.device.changeableValues.autoChangeoverActive}`);
+          payload.autoChangeoverActive = this.device.changeableValues.autoChangeoverActive;
+        }
         break;
       case 'Unknown':
-        if (this.platform.debugMode) {
-          this.platform.log.warn('autoChangeoverActive not sent for Unknown Thermostats');
-        }
+        this.platform.debug(`${this.device.deviceModel} Thermostats do not send autoChangeoverActive`);
         break;
       default:
-        if (this.platform.debugMode) {
-          this.platform.log.warn('set autoChangeoverActive');
-          this.platform.log.warn(this.device.deviceModel);
-        }
+        this.platform.debug(`Set autoChangeoverActive to ${this.device.changeableValues.autoChangeoverActive}, 
+        for ${this.device.deviceModel} Thermostats`);
         payload.autoChangeoverActive = this.device.changeableValues.autoChangeoverActive;
     }
 
@@ -535,14 +512,18 @@ export class Thermostats {
         );
     }
 
-    this.platform.log.debug('Thermostat %s pushChanges -', this.accessory.displayName, JSON.stringify(payload));
-    // Make the API request
-    await this.platform.axios.post(`${DeviceURL}/thermostats/${this.device.deviceID}`, payload, {
-      params: {
-        locationId: this.locationId,
-      },
-    });
-    // Refresh the status from the API
+    this.platform.log.debug(`Thermostat ${this.accessory.displayName} pushChanges - ${JSON.stringify(payload)}`);
+    // Attempt to make the API request
+    try {
+      await this.platform.axios.post(`${DeviceURL}/thermostats/${this.device.deviceID}`, payload, {
+        params: {
+          locationId: this.locationId,
+        },
+      });
+    } catch (e) {
+      // logged within post call above
+      this.apiError(e);
+    }
     await this.refreshStatus();
   }
 
@@ -550,12 +531,9 @@ export class Thermostats {
    * Pushes the requested changes for Room Priority to the Honeywell API
    */
   async pushRoomChanges() {
-    this.platform.log.debug(
-      'Thermostat Room Priority %s Current Room: %s, Changing Room: %s',
-      this.accessory.displayName,
-      JSON.stringify(this.roompriority.currentPriority.selectedRooms),
-      `[${this.device.inBuiltSensorState!.roomId}]`,
-    );
+    this.platform.debug(`Thermostat Room Priority for ${this.accessory.displayName}
+     Current Room: ${JSON.stringify(this.roompriority.currentPriority.selectedRooms)},
+     Changing Room: [${this.device.inBuiltSensorState!.roomId}]`);
     if (`[${this.device.inBuiltSensorState!.roomId}]` !== `[${this.roompriority.currentPriority.selectedRooms}]`) {
       const payload = {
         currentPriority: {
@@ -599,8 +577,7 @@ export class Thermostats {
             this.platform.config.options.roompriority.priorityType,
           );
         }
-        this.platform.log.debug('Thermostat %s pushRoomChanges -', this.accessory.displayName, JSON.stringify(payload));
-
+        this.platform.debug(`Thermostat ${this.accessory.displayName} pushRoomChanges - ${JSON.stringify(payload)}`);
         // Make the API request
         await this.platform.axios.put(`${DeviceURL}/thermostats/${this.device.deviceID}/priority`, payload, {
           params: {
@@ -691,7 +668,7 @@ export class Thermostats {
   }
 
   private setTargetHeatingCoolingState(value: CharacteristicValue) {
-    this.platform.log.debug('Thermostat %s -', this.accessory.displayName, 'Set TargetHeatingCoolingState:', value);
+    this.platform.debug(`Thermostat ${this.accessory.displayName}, Set TargetHeatingCoolingState: ${value}`);
 
     this.TargetHeatingCoolingState = value;
 
@@ -711,25 +688,25 @@ export class Thermostats {
   }
 
   private setHeatingThresholdTemperature(value: CharacteristicValue) {
-    this.platform.log.debug('Thermostat %s -', this.accessory.displayName, 'Set HeatingThresholdTemperature:', value);
+    this.platform.debug(`Thermostat ${this.accessory.displayName}, Set HeatingThresholdTemperature: ${value}`);
     this.HeatingThresholdTemperature = value;
     this.doThermostatUpdate.next();
   }
 
   private setCoolingThresholdTemperature(value: CharacteristicValue) {
-    this.platform.log.debug('Thermostat %s -', this.accessory.displayName, 'Set CoolingThresholdTemperature:', value);
+    this.platform.debug(`Thermostat ${this.accessory.displayName}, Set CoolingThresholdTemperature: ${value}`);
     this.CoolingThresholdTemperature = value;
     this.doThermostatUpdate.next();
   }
 
   private setTargetTemperature(value: CharacteristicValue) {
-    this.platform.log.debug('Thermostat %s -', this.accessory.displayName, 'Set TargetTemperature:', value);
+    this.platform.debug(`Thermostat ${this.accessory.displayName}, Set TargetTemperature: ${value}`);
     this.TargetTemperature = value;
     this.doThermostatUpdate.next();
   }
 
   private setTemperatureDisplayUnits(value: CharacteristicValue) {
-    this.platform.log.debug('Thermostat %s -', this.accessory.displayName, 'Set TemperatureDisplayUnits:', value);
+    this.platform.debug(`Thermostat ${this.accessory.displayName}, Set TemperatureDisplayUnits: ${value}`);
     this.platform.log.warn('Changing the Hardware Display Units from HomeKit is not supported.');
 
     // change the temp units back to the one the Honeywell API said the thermostat was set to
@@ -772,14 +749,7 @@ export class Thermostats {
       mode: 'Auto', // default to Auto
     };
     if (this.device.settings?.fan && !this.platform.config.options?.thermostat?.hide_fan) {
-      this.platform.log.debug(
-        'Thermostat %s -',
-        this.accessory.displayName,
-        'TargetFanState',
-        this.TargetFanState,
-        'Active',
-        this.Active,
-      );
+      this.platform.debug(`Thermostat ${this.accessory.displayName}, TargetFanState: ${this.TargetFanState}, Active: ${this.Active}`);
 
       if (this.TargetFanState === this.platform.Characteristic.TargetFanState.AUTO) {
         payload = {
@@ -807,7 +777,7 @@ export class Thermostats {
         'to Honeywell API. Fan Mode:',
         payload.mode,
       );
-      this.platform.log.debug('Thermostat %s Fan Mode -', this.accessory.displayName, JSON.stringify(payload));
+      this.platform.debug(`Thermostat ${this.accessory.displayName}, Fan Mode: ${JSON.stringify(payload)}`);
 
       // Make the API request
       await this.platform.axios.post(`${DeviceURL}/thermostats/${this.device.deviceID}/fan`, payload, {
@@ -824,19 +794,19 @@ export class Thermostats {
    * Updates the status for each of the HomeKit Characteristics
    */
   private setActive(value: CharacteristicValue) {
-    this.platform.log.debug('Thermostat %s -', this.accessory.displayName, 'Set Active State:', value);
+    this.platform.debug(`Thermostat ${this.accessory.displayName}, Set Active: ${value}`);
     this.Active = value;
     this.doFanUpdate.next();
   }
 
   private setTargetFanState(value: CharacteristicValue) {
-    this.platform.log.debug('Thermostat %s -', this.accessory.displayName, 'Set Target Fan State:', value);
+    this.platform.debug(`Thermostat ${this.accessory.displayName}, Set TargetFanState: ${value}`);
     this.TargetFanState = value;
     this.doFanUpdate.next();
   }
 
   private TargetState() {
-    this.platform.log.debug('Thermostat %s -', this.accessory.displayName, this.device.allowedModes);
+    this.platform.debug(`Thermostat ${this.accessory.displayName}, allowedModes: ${this.device.allowedModes}`);
 
     const TargetState = [4];
     TargetState.pop();
@@ -852,12 +822,7 @@ export class Thermostats {
     if (this.device.allowedModes.includes('Auto')) {
       TargetState.push(this.platform.Characteristic.TargetHeatingCoolingState.AUTO);
     }
-    this.platform.log.debug(
-      'Thermostat %s -',
-      this.accessory.displayName,
-      'Only Show These Modes:',
-      JSON.stringify(TargetState),
-    );
+    this.platform.debug(`Thermostat ${this.accessory.displayName}, Only Show These Modes: ${JSON.stringify(TargetState)}`);
     return TargetState;
   }
 }